+++ ZoneSharing/App/Main/Delegates/SceneDelegate.swift
@@ -23,13 +23,13 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 
     func windowScene(_ windowScene: UIWindowScene, userDidAcceptCloudKitShareWith cloudKitShareMetadata: CKShare.Metadata) {
-        guard cloudKitShareMetadata.containerIdentifier == Config.containerIdentifier else {
+        guard cloudKitShareMetadata.containerIdentifier == CKWorld.containerIdentifier else {
             print("Shared container identifier \(cloudKitShareMetadata.containerIdentifier) did not match known identifier.")
             return
         }
 
         // Create an operation to accept the share, running in the app's CKContainer.
-        let container = CKContainer(identifier: Config.containerIdentifier)
+        let container = CKContainer(identifier: CKWorld.containerIdentifier)
         let operation = CKAcceptSharesOperation(shareMetadatas: [cloudKitShareMetadata])
 
         debugPrint("Accepting CloudKit Share with metadata: \(cloudKitShareMetadata)")
