+++ ZoneSharing/App/Main/Helpers/Config.swift
@@ -1,9 +1,9 @@
 //
-//  Config.swift
+//  CKWorld.swift
 //  (cloudkit-samples) Zone Sharing
 //
 
-enum Config {
+enum CKWorld {
     /// iCloud container identifier.
     /// Update this if you wish to use your own iCloud container.
     static let containerIdentifier = "iCloud.com.apple.samples.cloudkit.zonesharing"
