+++ ZoneSharing/CloudKit/Creator.swift
@@ -8,11 +8,7 @@
 import Foundation
 import CloudKit
 
-struct CKWorld {
-    
-}
-
-@Observable @MainActor public class Creator {
+@Observable public class Creator {
     var time = [UUID: Socialite]()
     var currentSocialiteID: UUID?
     
@@ -34,3 +30,27 @@ struct CKWorld {
 private struct InvalidTime: Error {
     var errorDescription: String? { "Invalid or missing CKShare" }
 }
+
+private actor Airplane {
+    public let id = UUID()
+    var flight: CKFlight?
+    var passengers = Set<Socialite>()
+
+    var gate: String? {
+        flight.getGate(from airport: CKAirport)
+    }
+
+    init(flight: CKFlight? = nil) {
+        self.id = flight?.flightID ?? UUID().uuidString
+        self.flight = flight
+    }
+
+    func board(_ passenger: T) {
+        passengers.insert(passenger)
+    }
+
+    func deplane(_ passenger: T) {
+        passengers.remove(passenger)
+    }
+}
+
